例外処理を追加
