--- src/main/java/com/movie/movie/MovieControllerAdvice.java
@@ -1,25 +1,0 @@
-package com.movie.movie;
-
-import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-import java.time.ZonedDateTime;
-import java.util.Map;
-
-@ControllerAdvice
-public class MovieControllerAdvice {
-    @ExceptionHandler(value = MovieNotFoundException.class)
-    public ResponseEntity<Map<String, String>> handleMovieNotFoundException(
-            MovieNotFoundException e, HttpServletRequest request) {
-        Map<String, String> body = Map.of(
-                "timestamp", ZonedDateTime.now().toString(),
-                "status", String.valueOf(HttpStatus.NOT_FOUND.value()),
-                "error", HttpStatus.NOT_FOUND.getReasonPhrase(),
-                "message", e.getMessage(),
-                "path", request.getRequestURI());
-        return new ResponseEntity<>(body, HttpStatus.NOT_FOUND);
-    }
-}
