--- src/main/java/com/movie/movie/MovieNotFoundException.java
@@ -1,7 +1,0 @@
-package com.movie.movie;
-
-public class MovieNotFoundException extends RuntimeException {
-    public MovieNotFoundException(String message) {
-        super(message);
-    }
-}
