+++ src/main/java/com/movie/movie/exception/InvalidYearException.java
@@ -1,0 +1,8 @@
+package com.movie.movie.exception;
+
+    public class InvalidYearException extends RuntimeException {
+        public InvalidYearException(String message) {
+            super(message);
+        }
+    }
+
