+++ src/main/java/com/movie/movie/exception/MovieNotFoundException.java
@@ -1,0 +1,8 @@
+package com.movie.movie.exception;
+
+public class MovieNotFoundException extends RuntimeException {
+    public MovieNotFoundException(String message) {
+        super(message);
+    }
+}
+
